+++ plone/app/contenttypes/migration/browser.py
@@ -37,7 +37,6 @@ from zope.component import getGlobalSiteManager
 
 from plone.app.contenttypes.content import (
     Document,
-    Event,
     File,
     Folder,
     Image,
@@ -58,7 +57,6 @@ class FixBaseClasses(BrowserView):
         out = ""
         portal_types = [
             ('Document', Document),
-            ('Event', Event),
             ('File', File),
             ('Folder', Folder),
             ('Image', Image),