+++ plone/app/contenttypes/testing.py
@@ -7,7 +7,7 @@ from plone.app.testing import FunctionalTesting
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
 from plone.app.testing import login
-
+from plone.app.event.testing import PAEvent_FIXTURE
 from plone.testing import z2
 
 from zope.configuration import xmlconfig
@@ -15,7 +15,7 @@ from zope.configuration import xmlconfig
 
 class PloneAppContenttypes(PloneSandboxLayer):
 
-    defaultBases = (PLONE_FIXTURE,)
+    defaultBases = (PAEvent_FIXTURE, PLONE_FIXTURE)
 
     def setUpZope(self, app, configurationContext):
         import plone.app.contenttypes